Applied gimp-simon-980331-0. Cosmetic change to the Layers&Channels-dialog.
I will hopefully finish my changes to the dialog tonight and will upload a
new version of my patch. Could you (the core developers) please make up your
mind whether to include it in 1.0 or not...

The Blend-Tool dialog now changes the repeat_mode_menu to insensitive when
the Gradient-Type is one of the Shapeburst-types.


--Sven


git-svn-id: http://svn.gnome.org/svn/gimp/trunk@414 f57ce169-c725-0410-9505-c4955b3f21d2