Sun Jul 12 19:00:15 EDT 1998 Michael K. Johnson <johnsonm@redhat.com>
* app/tile_swap.c
* app/tile_swap.h
* app/tile_manager.c
* app/tile_manager.h
* app/pixel_region.c
* app/pixel_region.h: asynchronous swapin on systems with pthreads.
This version is not at all tuned, and the only interface which makes use
of it now is pixel_region_{g,s}et_{row,col}.  Other functions which know
ahead of time the area that they will be needing can request that it be
asynchronously swapped in via the pixel_region_get_async() function.

Compiles and survives basic testing.


git-svn-id: http://svn.gnome.org/svn/gimp/trunk@842 f57ce169-c725-0410-9505-c4955b3f21d2